Renames files, variables in prep for migration and updates to latest library versions.
