+++ README.md
@@ -1,24 +1,6 @@
-# data-providers
+# Exposing data to watch face complications on Wear OS
 
-This folder contains the source code for a Google I/O 2016 codelab.
+Learn how to share you data with complications, so users can easily enter back into your app and
+get the latest information from your app.
 
-### License
-
-```
-Copyright 2016 Google, Inc.
-
-Licensed to the Apache Software Foundation (ASF) under one or more contributor
-license agreements. See the NOTICE file distributed with this work for
-additional information regarding copyright ownership. The ASF licenses this
-file to you under the Apache License, Version 2.0 (the "License"); you may not
-use this file except in compliance with the License. You may obtain a copy of
-the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-License for the specific language governing permissions and limitations under
-the License.
-```
+To see the actual code lab, [click here](https://developer.android.com/codelabs/data-providers).
