+++ base/src/main/AndroidManifest.xml
@@ -13,8 +13,10 @@
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
   -->
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.android.wearable.complicationsdataprovider">
+<manifest
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    package="com.example.android.wearable.complicationsdatasource">
 
     <uses-feature android:name="android.hardware.type.watch" />
 
@@ -23,12 +25,17 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@android:style/Theme.DeviceDefault">
+        android:theme="@android:style/Theme.DeviceDefault"
+        tools:ignore="AllowBackup">
 
+        <meta-data
+            android:name="com.google.android.wearable.standalone"
+            android:value="true" />
         <service
-            android:name=".CustomComplicationProviderService"
+            android:name="com.example.android.wearable.complicationsdatasource.CustomComplicationDataSourceService"
+            android:exported="true"
             android:icon="@drawable/ic_watch_white"
-            android:label="@string/complications_provider_service_label"
+            android:label="@string/complications_data_sources_service_label"
             android:permission="com.google.android.wearable.permission.BIND_COMPLICATION_PROVIDER">
             <intent-filter>
                 <action android:name="android.support.wearable.complications.ACTION_COMPLICATION_UPDATE_REQUEST" />
@@ -42,7 +49,7 @@
                 android:name="android.support.wearable.complications.UPDATE_PERIOD_SECONDS"
                 android:value="600" />
         </service>
-        <receiver android:name=".ComplicationTapBroadcastReceiver" />
+        <receiver android:name="com.example.android.wearable.complicationsdatasource.ComplicationTapBroadcastReceiver" />
     </application>
 
 </manifest>
