+++ build.gradle
@@ -18,30 +18,30 @@ buildscript {
     ext {
 
         // App version
-        versionName = '1.0.0'
+        versionName = '1.1.0'
         versionCode = 1
 
         // SDK and tools
-        compileSdkVersion = 28
-        minSdkVersion = 23
-        targetSdkVersion = 28
+        compileSdkVersion = 31
+        minSdkVersion = 26
+        targetSdkVersion = 30
 
         // App dependencies
-        kotlinVersion = '1.4.10'
-        coreKtxVersion = '1.3.2'
+        kotlinVersion = '1.5.21'
+        coreKtxVersion = '1.6.0'
         androidxPercentlayoutVersion = '1.0.0'
-        androidxRecyclerviewVersion = '1.1.0'
+        androidxRecyclerviewVersion = '1.2.1'
         supportLibraryVersion = '28.0.0'
-        googlePlayServicesClientVersion = '17.0.0'
+        googlePlayServicesClientVersion = '17.1.0'
         wearableSupportLibraryVersion = '2.8.1'
     }
 
     repositories {
-        jcenter()
         google()
+        mavenCentral()
     }
     dependencies {
-        classpath "com.android.tools.build:gradle:4.1.0"
+        classpath 'com.android.tools.build:gradle:7.0.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
 
         // NOTE: Do not place your application dependencies here; they belong
@@ -53,7 +53,6 @@ allprojects {
     repositories {
         google()
         mavenCentral()
-        jcenter()
     }
 }
